requirements/production.txt
django
djangorestframework

# openid-connect mit JWT
djangorestframework-simplejwt[crypto]

# OpenAPI - drf-spectacular
drf-spectacular

gunicorn
